# api-server/openapi.yaml
openapi: 3.1.0
info:
  title: Humanizer API
  description: AI writing detection and humanization API. Detects 29 patterns of AI-generated text and provides actionable suggestions. Supports English (en) and Swedish (sv) via the optional locale field.
  version: 2.2.0
  contact:
    name: Brandon Wise
    url: https://github.com/brandonwise/humanizer

servers:
  - url: https://humanizer.example.com/api
    description: Production server

components:
  schemas:
    LocaleEnum:
      type: string
      enum: [en, sv]
      default: en
      description: Language locale. Use "sv" for Swedish text (Swedish vocabulary tiers, LIX readability, Swedish abbreviation handling).

paths:
  /score:
    post:
      operationId: scoreText
      summary: Quick AI score
      description: Returns a score from 0-100 indicating how AI-like the text appears. 0-25 human, 26-50 light AI, 51-75 moderate, 76-100 heavy AI.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - text
              properties:
                text:
                  type: string
                  description: Text to analyze
                locale:
                  $ref: '#/components/schemas/LocaleEnum'
      responses:
        '200':
          description: Score result
          content:
            application/json:
              schema:
                type: object
                properties:
                  score:
                    type: integer
                    minimum: 0
                    maximum: 100
                  badge:
                    type: string
                    enum: ['🟢', '🟡', '🟠', '🔴']
                  interpretation:
                    type: string
                  locale:
                    type: string

  /analyze:
    post:
      operationId: analyzeText
      summary: Full AI analysis
      description: Complete analysis with pattern matches, category breakdown, and statistical analysis. Pass locale "sv" for Swedish text.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - text
              properties:
                text:
                  type: string
                  description: Text to analyze
                verbose:
                  type: boolean
                  default: false
                  description: Include all pattern matches
                locale:
                  $ref: '#/components/schemas/LocaleEnum'
      responses:
        '200':
          description: Analysis result
          content:
            application/json:
              schema:
                type: object
                properties:
                  score:
                    type: integer
                  patternScore:
                    type: integer
                  uniformityScore:
                    type: integer
                  categories:
                    type: object
                  stats:
                    type: object
                    description: Includes fleschKincaid (en) or lix (sv)
                  findings:
                    type: array
                    items:
                      type: object

  /humanize:
    post:
      operationId: humanizeText
      summary: Get humanization suggestions
      description: Returns suggestions grouped by priority to make text sound more human, with optional auto-fix. Pass locale "sv" for Swedish text — guidance strings and autofix rules will be Swedish-native.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - text
              properties:
                text:
                  type: string
                  description: Text to humanize
                autofix:
                  type: boolean
                  default: false
                  description: Apply safe mechanical fixes
                locale:
                  $ref: '#/components/schemas/LocaleEnum'
      responses:
        '200':
          description: Humanization suggestions
          content:
            application/json:
              schema:
                type: object
                properties:
                  critical:
                    type: array
                    items:
                      type: object
                  important:
                    type: array
                    items:
                      type: object
                  minor:
                    type: array
                    items:
                      type: object
                  guidance:
                    type: array
                    items:
                      type: string
                  styleTips:
                    type: array
                    items:
                      type: object
                  autofix:
                    type: object
                    properties:
                      text:
                        type: string
                      fixes:
                        type: array
                        items:
                          type: string

  /stats:
    post:
      operationId: statsText
      summary: Statistical analysis only
      description: Returns burstiness, type-token ratio, sentence variation, and readability metrics. For Swedish (locale "sv"), returns LIX instead of Flesch-Kincaid.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - text
              properties:
                text:
                  type: string
                locale:
                  $ref: '#/components/schemas/LocaleEnum'
      responses:
        '200':
          description: Statistical analysis
          content:
            application/json:
              schema:
                type: object
                properties:
                  wordCount:
                    type: integer
                  sentenceCount:
                    type: integer
                  burstiness:
                    type: number
                  typeTokenRatio:
                    type: number
                  sentenceLengthVariation:
                    type: number
                  trigramRepetition:
                    type: number
                  fleschKincaid:
                    type: number
                    nullable: true
                    description: English readability grade (null when locale is sv)
                  lix:
                    type: number
                    nullable: true
                    description: Swedish LIX readability score (null when locale is en)